adding set methods

+++ src/Repository/Write/IAsyncWriteRepository.cs
@@ -7,30 +7,31 @@ namespace eQuantic.Core.Data.Repository.Write
 {
     public interface IAsyncWriteRepository<TUnitOfWork, TEntity, TKey> : IAsyncWriteRepository<TEntity, TKey>, IAsyncRepository<TUnitOfWork>
         where TUnitOfWork : IUnitOfWork
-    where TEntity : class, IEntity, new () { }
+    where TEntity : class, IEntity, new()
+    { }
 
     public interface IAsyncWriteRepository<TEntity, TKey> : IAsyncRepository
-    where TEntity : class, IEntity, new ()
+    where TEntity : class, IEntity, new()
     {
         /// <summary>
         /// Add item into repository
         /// </summary>
         /// <param name="item">Item to add to repository</param>
-        Task AddAsync (TEntity item);
+        Task AddAsync(TEntity item);
 
         /// <summary>
         /// Delete filtered elements of type TEntity in repository
         /// </summary>
         /// <param name="filter"></param>
         /// <returns></returns>
-        Task<int> DeleteManyAsync (Expression<Func<TEntity, bool>> filter);
+        Task<long> DeleteManyAsync(Expression<Func<TEntity, bool>> filter);
 
         /// <summary>
         /// Delete specified elements of type TEntity in repository
         /// </summary>
         /// <param name="specification"></param>
         /// <returns></returns>
-        Task<int> DeleteManyAsync (ISpecification<TEntity> specification);
+        Task<long> DeleteManyAsync(ISpecification<TEntity> specification);
 
         /// <summary>
         /// Sets modified entity into the repository. When calling Commit() method in UnitOfWork
@@ -38,19 +39,19 @@ namespace eQuantic.Core.Data.Repository.Write
         /// </summary>
         /// <param name="persisted">The persisted item</param>
         /// <param name="current">The current item</param>
-        Task MergeAsync (TEntity persisted, TEntity current);
+        Task MergeAsync(TEntity persisted, TEntity current);
 
         /// <summary>
         /// Set item as modified
         /// </summary>
         /// <param name="item">Item to modify</param>
-        Task ModifyAsync (TEntity item);
+        Task ModifyAsync(TEntity item);
 
         /// <summary>
         /// Delete item
         /// </summary>
         /// <param name="item">Item to delete</param>
-        Task RemoveAsync (TEntity item);
+        Task RemoveAsync(TEntity item);
 
         /// <summary>
         /// Update filtered elements of type TEntity in repository
@@ -58,7 +59,7 @@ namespace eQuantic.Core.Data.Repository.Write
         /// <param name="filter"></param>
         /// <param name="updateFactory"></param>
         /// <returns></returns>
-        Task<int> UpdateManyAsync (Expression<Func<TEntity, bool>> filter, Expression<Func<TEntity, TEntity>> updateFactory);
+        Task<long> UpdateManyAsync(Expression<Func<TEntity, bool>> filter, Expression<Func<TEntity, TEntity>> updateFactory);
 
         /// <summary>
         /// Update specified elements of type TEntity in repository
@@ -66,6 +67,6 @@ namespace eQuantic.Core.Data.Repository.Write
         /// <param name="specification"></param>
         /// <param name="updateFactory"></param>
         /// <returns></returns>
-        Task<int> UpdateManyAsync (ISpecification<TEntity> specification, Expression<Func<TEntity, TEntity>> updateFactory);
+        Task<long> UpdateManyAsync(ISpecification<TEntity> specification, Expression<Func<TEntity, TEntity>> updateFactory);
     }
 }
